jd02_06(A,B) jd02_07(A,B)
