+++ src/by/it/bolotko/jd02_06/calc/log.txt
@@ -17,3 +17,5 @@
 26.04.2019 05:42:37 : 32.0
 26.04.2019 05:42:39 : 333.0
 26.04.2019 05:42:44 : 4440000.0
+29.04.2019 02:49:07 : 554.0
+29.04.2019 02:49:15 : {A={{25.0, 21.0}, {36.0, 31.0}}, F=554.0}
