+++ src/by/it/bolotko/jd02_06/calc/vars.txt
@@ -1,1 +1,2 @@
 A={{25.0, 21.0}, {36.0, 31.0}}
+F=554.0
