+++ src/by/it/bolotko/jd02_07/parkings.dtd
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+        <!ELEMENT parkings (parking)+>
+        <!ELEMENT parking (first_name,last_name,year_of_birth,drivers_license_number,category,cars)>
+        <!ELEMENT first_name (#PCDATA)>
+        <!ELEMENT last_name (#PCDATA)>
+        <!ELEMENT year_of_birth (#PCDATA)>
+        <!ELEMENT drivers_license_number (#PCDATA)>
+        <!ELEMENT category (#PCDATA)>
+        <!ELEMENT cars (car,prices)+>
+        <!ELEMENT car (car_model,car_type,numbers)
+        <!ELEMENT car_model (#PCDATA)>
+        <!ELEMENT car_type (#PCDATA)>
+        <!ELEMENT numbers (#PCDATA)>
+        <!ELEMENT prices (price)+>
+        <!ELEMENT price (number_of_cars,total_price)>
+        <!ELEMENT number_of_cars (#PCDATA)>
+        <!ELEMENT total_price (#PCDATA)>
+        <!ATTLIST total_price
+                currency PCDATA #REQUIRED
+                >
